+++ ft_strncmp.c
@@ -6,13 +6,15 @@
 /*   By: maemaldo <maemaldo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/06 16:12:18 by maemaldo          #+#    #+#             */
-/*   Updated: 2023/11/06 17:15:32 by maemaldo         ###   ########.fr       */
+/*   Updated: 2023/11/08 10:28:50 by maemaldo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include<stdlib.h>
+
 int ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (s1[i] && s2[i] && i <= n)
